Add files via upload
As of RED 8.0.2, template functionality does not interact with the metadata repository to specify the TPT write operator (LOAD, UPDATE, STREAM).   As a result, the TLOGVIEW output in this template will only properly report a LOAD job back to the user.   This script only works on the Windows Operating System.